+++ LiteOrm.Demo/Demos/MySqlBulkInsertProvider.cs
@@ -1,13 +1,22 @@
-﻿using LiteOrm.Common;
-using MySqlConnector;
+﻿using MySqlConnector;
 using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace LiteOrm.Demo.Demos
 {
-    [BulkProvider(typeof(MySqlConnection))]
+    /// <summary>
+    /// MySQL 批量插入提供程序。
+    /// 使用方式：直接设置到 <see cref="LiteOrm.SqlBuilder"/> 的 <c>BulkProvider</c> 属性，例如：
+    /// <code>
+    /// LiteOrm.SqlBuilderFactory.Instance.GetSqlBuilder(typeof(MySqlConnection)).BulkProvider = new MySqlBulkCopyProvider();
+    /// 或
+    /// MySqlBuilder.Instance.BulkProvider = new MySqlBulkCopyProvider();
+    /// </code>
+    /// </summary>
     public class MySqlBulkCopyProvider : IBulkProvider
     {
         public int BulkInsert(DataTable dt, IDbConnection dbConnection, IDbTransaction transaction)