+++ LiteOrm/DAO/ObjectDAO.cs
@@ -479,7 +479,7 @@ namespace LiteOrm
         /// <param name="values">要插入的实体对象集合。</param>
         public virtual void BatchInsert(IEnumerable<T> values)
         {
-            var provider = BulkFactory?.GetProvider(GetProviderType()!);
+            var provider = SqlBuilder.BulkProvider;
             var insertableColumns = TableDefinition.Columns.Where(column => !column.IsIdentity && column.Mode.CanInsert()).ToArray();
             var daoContext = GetDaoContext();
             if (provider is not null)
@@ -941,7 +941,7 @@ namespace LiteOrm
         /// <returns>表示异步操作的任务。</returns>
         public async virtual Task BatchInsertAsync(IEnumerable<T> values, CancellationToken cancellationToken = default)
         {
-            var provider = BulkFactory?.GetProvider(GetProviderType()!);
+            var provider = SqlBuilder.BulkProvider;
             var insertableColumns = TableDefinition.Columns.Where(column => !column.IsIdentity && column.Mode.CanInsert()).ToArray();
             var daoContext = await GetDaoContextAsync(cancellationToken).ConfigureAwait(false);
             if (provider is not null)